+++ src/03-modules/review-add/ReviewAdd.vue
@@ -4,9 +4,8 @@
 import AHeading from '../../atoms/heading/Heading.vue'
 import AParagraph from '../../atoms/paragraph/Paragraph.vue'
 import AButton from '../../atoms/button/Button.vue'
-import AInput from '../../atoms/input/Input.vue'
+import AInput from '../../molecules/input/Input.vue'
 import ATextarea from '../../atoms/textarea/Textarea.vue'
-import AInputWrapper from '../../molecules/input-wrapper/InputWrapper.vue'
 import AFieldset from '../../atoms/fieldset/Fieldset.vue'
 
 import AReviewAdd from './ReviewAdd.js'
@@ -18,7 +17,6 @@ export default {
     AParagraph,
     AButton,
     AInput,
-    AInputWrapper,
     AFieldset,
     ATextarea
   },